requirements.txt
appdirs==1.4.4
attrs==20.1.0
Automat==20.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
Bootstrap-Flask==1.3.0
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
configparser==5.0.0
constantly==15.1.0
crayons==0.4.0
cryptography==3.1
cssselect==1.1.0
distlib==0.3.1
docopt==0.6.2
filelock==3.0.12
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-Views==0.2.1
Flask-WTF==0.14.2
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
inexactsearch==1.0.2
itemadapter==0.1.0
itemloaders==1.0.2
itsdangerous==1.1.0
Jinja2==2.11.0
jmespath==0.10.0
Js2Py==0.70
lxml==4.5.2
MarkupSafe==1.1.1
packaging==20.4
parsel==1.6.0
pipwin==0.5.0
Protego==0.1.16
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyjsparser==2.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPrind==2.11.2
pyquery==1.4.1
pySmartDL==1.3.3
pyspellchecker==0.5.5
pytz==2020.1
queuelib==1.5.0
rauth==0.7.3
regex==2020.6.8
requests==2.24.0
Scrapy==2.3.0
service-identity==18.1.0
silpa-common==0.3
six==1.15.0
soundex==1.1.3
soupsieve==2.0.1
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.3
Twisted==20.3.0
tzlocal==2.1
urllib3==1.24.3
URLObject==2.4.3
virtualenv==20.0.4
w3lib==1.22.0
webdriver-manager==3.2.1
Werkzeug==0.16.1
WTForms==2.2.1
xvfbwrapper==0.2.9
zipp==3.1.0
zope.interface==5.1.0
